requirements.txt
streamlit
InstructorEmbedding
ctransformers
transformers
langchain
streamlit
pypdfium2
librosa
pyyaml